Page loaders for shows and wrestlers
These add `/shows/<id_array>` and `wrestlers/<id>` page loaders.

The main page has been updated so that the list of wrestlers ordered by
matches in the DB, now contains links to the respective wrestlers page
loader. Where a wrestler has multiple names, those are now displayed in
order of the number of times that name was used (this is done with a
new macro in `shows.db.sh`).

The wrestlers loader displays every match in the DB that wrestler has had, in
chronological order, grouped by show, with a link to the shows loader page
for each, alongside general stats like number of matches, number of
promotions. Also links to the original CM profile.

The shows loader displays the info in the DB for that show, along
with links to the original CM page.