Actually generate page loader pages in build
While I added page loaders that work in live dev server, to actually get
them as part of our built static site, we need to list them in
`dynamicPaths`. This is done with a function that creates the DuckDB
database in the same way as the page loader, then queries it for the ID
numbers we need for each loader.

They are not yet listed in the sidebar: the `pages` option used to
configure this doesn't currently allow functions. Once the pages are
being actually fully generated rather than dynamically loaded from DuckDB,
we could turn on the search box, this should allow you to discover
anything we've generated a page for, and related pages.

- Add `@duckdb/node-api` dependency for config
- Add `dynamicPaths` to `observablehq.config.js` to enumerate page
  loader pages to generate in build